pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "object-detection-app"
version = "0.1.0"
description = "Application de détection d'objets en temps réel avec génération de descriptions."
authors = [{name = "Votre Nom", email = "votre.email@example.com"}]
dependencies = ["torch", "transformers", "opencv-python", "Pillow", "gradio"]
